Moved missions stuff into its own module.
Ruff fix

Ignore the positional arguments error.

Put the ignore in the right place.
