+++ pyproject.toml
@@ -91,6 +91,7 @@ ignore = [
     "PLR0912", # too many branches
     "PLR0913", # too many arguments
     "PLR0915", # too many statements
+	"PLR0917", # too many positional arguments
     "PLR2004", # magic value comparison
     "PLC0415", # import should be at top of file
 ]
